requirements.txt
3-1==1.0.0
absl-py==0.9.0
aioredis==1.3.1
amqp==2.6.0
appdirs==1.4.4
ar-virtualenv-api==2.0.0
asgiref==3.2.7
astroid==2.2.5
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.3.0
autobahn==20.7.1
Automat==20.2.0
awscli==1.18.109
backcall==0.1.0
beautifulsoup4==4.9.1
billiard==3.6.3.0
bleach==3.1.5
boto==2.49.0
boto3==1.14.32
botocore==1.17.32
cachetools==4.1.1
celery==4.4.6
certifi==2020.6.20
cffi==1.14.0
channels==2.4.0
channels-redis==3.1.0
chardet==3.0.4
click==7.1.2
cloudpickle==1.3.0
colorama==0.4.3
configparser==5.0.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
cycler==0.10.0
Cython==0.29.13
daphne==2.5.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.8
django-allauth==0.42.0
django-appconf==1.0.4
django-bootstrap-static==4.2.1
django-braces==1.14.0
django-ckeditor==5.9.0
django-cleanup==5.0.0
django-compressor==2.4
django-cors-headers==3.4.0
django-countries==6.1.2
django-crispy-forms==1.9.1
django-cropperjs==0.0.5
django-debug-toolbar==2.2
django-embed-video==1.3.3
django-filebrowser==3.13.1
django-filter==2.3.0
django-grappelli==2.14.2
django-haystack==3.0b2
django-imagekit==4.0.2
django-js-asset==1.2.2
django-mdeditor==0.1.18
django-memcached==0.1.2
django-registration-redux==2.8
django-rest-auth==0.9.5
django-rest-multiple-models==2.1.3
django-rest-swagger==2.2.0
django-storages==1.9.1
django-taggit==1.3.0
django-taggit-serializer==0.1.7
djangorestframework==3.11.0
docopt==0.6.2
docutils==0.15.2
drf-yasg==1.20.0
entrypoints==0.3
facebook-sdk==3.1.0
filelock==3.0.12
future==0.18.2
gl==0.2.36
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.19.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
gunicorn==20.0.4
h2==3.2.0
h5py==2.9.0
hiredis==1.1.0
hpack==3.0.0
httplib2==0.18.1
hyperframe==5.2.0
hyperlink==20.0.1
idna==2.9
importlib-metadata==1.6.1
incremental==17.5.0
inflection==0.5.1
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.15
itsdangerous==1.1.0
itypes==1.2.0
jedi==0.13.2
Jinja2==2.11.2
jmespath==0.10.0
jsmin==2.2.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
jwcrypto==0.7
Keras==2.4.3
kiwisolver==1.1.0
kombu==4.6.11
lazy-object-proxy==1.3.1
lxml==4.4.1
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
mistune==0.8.4
msgpack==1.0.0
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
numpy==1.16.2
oauthlib==3.1.0
openapi-codec==1.3.2
opengl-registry==0.1.0
packaging==20.4
pandocfilters==1.4.2
parso==0.3.4
pexpect==4.8.0
pickleshare==0.7.5
pilkit==2.0
Pillow==6.1.0
pipenv==2020.6.2
priority==1.3.0
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
psutil==5.5.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyglet==1.5.0
Pygments==2.6.1
PyHamcrest==2.0.2
PyJWT==2.0.0
pylint==2.3.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.16.0
PySDL2==0.9.4
python-dateutil==2.8.1
python-decouple==3.3
python-git==2018.2.1
python-twitter==3.5
python3-memcached==1.51
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
pyzmq==18.1.0
qtconsole==4.7.4
QtPy==1.9.0
rcssmin==1.0.6
requests==2.24.0
requests-oauthlib==1.3.0
rjsmin==1.1.0
rsa==4.5
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
runcmd==0.0.0
s3transfer==0.3.3
scipy==1.3.1
seaborn==0.10.1
Send2Trash==1.5.0
service-identity==18.1.0
simplejson==3.17.2
six==1.12.0
sklearn==0.0
soupsieve==2.0.1
sqlparse==0.3.1
stripe==2.50.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
Twisted==20.3.0
txaio==20.4.1
typed-ast==1.3.1
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
wcwidth==0.2.3
webencodings==0.5.1
websockets==8.1
Werkzeug==1.0.1
whitenoise==5.1.0
widgetsnbextension==3.5.1
wrapt==1.11.1
zipp==3.1.0
zope.interface==5.1.0
